[feat #2 #18] 로그인, 회원가입 기능 구현

+++ momentum-dao-be/src/main/java/com/dao/momentum/organization/employee/command/domain/aggregate/EmployeeRoles.java
@@ -15,9 +15,9 @@ import lombok.NoArgsConstructor;
 public class EmployeeRoles {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer employeeRolesId;
+    private Long employeeRolesId;
 
-    private Integer empId;
+    private Long empId;
 
     private Integer UserRoleId;
 }
